Improve planogram restore, tooltips, and type safety
- Add type casting for IDs in controllers for type safety
- Add null checks for user profile region/country in PlanogramsController
- Implement restorePlanogram method in PlanogramService
- Add "Restore" button and tooltips to planogramListView.vue
- Standardize "Actions" column header across list views
- Add consistent tooltip styling and theme config
- Enhance PlanogramRepository region/country param handling
- Register new PrimeVue icon components in main.ts
- Add SCSS for tooltips and minor style/import improvements
- Update submodules to latest dirty states

+++ PM-AdminApp/PM-AdminApp.Server/Controllers/Planner/ClusterController.cs
@@ -851,14 +851,14 @@ namespace PM_AdminApp.Server.Controllers.Planner
                     shelfId = 0;
                 }
 
-                newShelf.ClusterId = shelf.PlanogramId;
+                newShelf.ClusterId = (long)shelf.PlanogramId;
                 newShelf.ShelfTypeId = shelf.ShelfTypeId;
                 newShelf.Height = (short)shelf.Height;
                 newShelf.Width = (short)shelf.Width;
                 newShelf.PositionX = shelf.Position.x;
                 newShelf.PositionY = shelf.Position.y;
 
-                newShelf.PartId = shelf.PartId;
+                newShelf.PartId = (long)shelf.PartId;
                 newShelf.PartStatusId = shelf.StatusId ?? 0;
                 newShelf.Label = shelf.Label;
